--- interpretator/srcs/ft_stricut.c
@@ -1,13 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_stricut.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mmarti <marvin@42.fr>                      +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/11/29 16:22:49 by mmarti            #+#    #+#             */
-/*   Updated: 2019/11/29 16:22:50 by mmarti           ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libstr.h"
